Removing inline, more refactor out of math.ts
