+++ packages/math/point.ts
@@ -1,4 +1,4 @@
-import { degreesToRadians } from "./utils";
+import { degreesToRadians } from "./angle";
 import type {
   LocalPoint,
   GlobalPoint,
@@ -7,6 +7,7 @@ import type {
   Vector,
 } from "./types";
 import { PRECISION } from "./utils";
+import { vectorFromPoint, vectorScale } from "./vector";
 
 /**
  * Create a properly typed Point instance from the X and Y coordinates.
@@ -19,7 +20,6 @@ export function point<Point extends GlobalPoint | LocalPoint>(
   x: number,
   y: number,
 ): Point {
-  "inline";
   return [x, y] as Point;
 }
 
@@ -32,7 +32,6 @@ export function point<Point extends GlobalPoint | LocalPoint>(
 export function pointFromArray<Point extends GlobalPoint | LocalPoint>(
   numberArray: number[],
 ): Point | undefined {
-  "inline";
   return numberArray.length === 2
     ? point<Point>(numberArray[0], numberArray[1])
     : undefined;
@@ -47,7 +46,6 @@ export function pointFromArray<Point extends GlobalPoint | LocalPoint>(
 export function pointFromPair<Point extends GlobalPoint | LocalPoint>(
   pair: [number, number],
 ): Point {
-  "inline";
   return pair as Point;
 }
 
@@ -60,7 +58,6 @@ export function pointFromPair<Point extends GlobalPoint | LocalPoint>(
 export function pointFromVector<P extends GlobalPoint | LocalPoint>(
   v: Vector,
 ): P {
-  "inline";
   return v as unknown as P;
 }
 
@@ -71,7 +68,6 @@ export function pointFromVector<P extends GlobalPoint | LocalPoint>(
  * @returns TRUE if the provided value has the shape of a local or global point
  */
 export function isPoint(p: unknown): p is LocalPoint | GlobalPoint {
-  "inline";
   return (
     Array.isArray(p) &&
     p.length === 2 &&
@@ -94,7 +90,6 @@ export function pointsEqual<Point extends GlobalPoint | LocalPoint>(
   a: Point,
   b: Point,
 ): boolean {
-  "inline";
   const abs = Math.abs;
   return abs(a[0] - b[0]) < PRECISION && abs(a[1] - b[1]) < PRECISION;
 }
@@ -112,8 +107,6 @@ export function pointRotateRads<Point extends GlobalPoint | LocalPoint>(
   [cx, cy]: Point,
   angle: Radians,
 ): Point {
-  "inline";
-
   return point(
     (x - cx) * Math.cos(angle) - (y - cy) * Math.sin(angle) + cx,
     (x - cx) * Math.sin(angle) + (y - cy) * Math.cos(angle) + cy,
@@ -133,8 +126,6 @@ export function pointRotateDegs<Point extends GlobalPoint | LocalPoint>(
   center: Point,
   angle: Degrees,
 ): Point {
-  "inline";
-
   return pointRotateRads(point, center, degreesToRadians(angle));
 }
 
@@ -155,7 +146,6 @@ export function pointTranslate<
   From extends GlobalPoint | LocalPoint,
   To extends GlobalPoint | LocalPoint,
 >(p: From, v: Vector = [0, 0] as Vector): To {
-  "inline";
   return point(p[0] + v[0], p[1] + v[1]);
 }
 
@@ -167,8 +157,14 @@ export function pointTranslate<
  * @returns
  */
 export function pointCenter<P extends LocalPoint | GlobalPoint>(a: P, b: P): P {
-  "inline";
   return point((a[0] + b[0]) / 2, (a[1] + b[1]) / 2);
+}
+
+export function pointAdd<Point extends LocalPoint | GlobalPoint>(
+  a: Point,
+  b: Point,
+): Point {
+  return point(a[0] + b[0], a[1] + b[1]);
 }
 
 /**
@@ -182,7 +178,6 @@ export function pointDistance<P extends LocalPoint | GlobalPoint>(
   a: P,
   b: P,
 ): number {
-  "inline";
   return Math.hypot(b[0] - a[0], b[1] - a[1]);
 }
 
@@ -199,69 +194,18 @@ export function pointDistanceSq<P extends LocalPoint | GlobalPoint>(
   a: P,
   b: P,
 ): number {
-  "inline";
   return Math.hypot(b[0] - a[0], b[1] - a[1]);
 }
 
-// const topPointFirst = <Point extends LocalPoint | GlobalPoint>(
-//   line: Line<Point>,
-// ) => {
-//   return line[1][1] > line[0][1] ? line : [line[1], line[0]];
-// };
-
-// export const pointLeftofLine = <Point extends LocalPoint | GlobalPoint>(
-//   point: Point,
-//   line: Line<Point>,
-// ) => {
-//   const t = topPointFirst(line);
-//   return (
-//     vectorCross(vectorFromPoint(point, t[0]), vectorFromPoint(t[1], t[0])) < 0
-//   );
-// };
-
-// export const pointRightofLine = <Point extends LocalPoint | GlobalPoint>(
-//   point: Point,
-//   line: Line<Point>,
-// ) => {
-//   const t = topPointFirst(line);
-//   return (
-//     vectorCross(vectorFromPoint(point, t[0]), vectorFromPoint(t[1], t[0])) > 0
-//   );
-// };
-
-// export const pointInBezierEquation = <Point extends LocalPoint | GlobalPoint>(
-//   p0: Point,
-//   p1: Point,
-//   p2: Point,
-//   p3: Point,
-//   [mx, my]: Point,
-//   lineThreshold: number,
-// ) => {
-//   // B(t) = p0 * (1-t)^3 + 3p1 * t * (1-t)^2 + 3p2 * t^2 * (1-t) + p3 * t^3
-//   const equation = (t: number, idx: number) =>
-//     Math.pow(1 - t, 3) * p3[idx] +
-//     3 * t * Math.pow(1 - t, 2) * p2[idx] +
-//     3 * Math.pow(t, 2) * (1 - t) * p1[idx] +
-//     p0[idx] * Math.pow(t, 3);
-
-//   const lineSegmentPoints: Point[] = [];
-//   let t = 0;
-//   while (t <= 1.0) {
-//     const tx = equation(t, 0);
-//     const ty = equation(t, 1);
-
-//     const diff = Math.sqrt(Math.pow(tx - mx, 2) + Math.pow(ty - my, 2));
-
-//     if (diff < lineThreshold) {
-//       return true;
-//     }
-
-//     lineSegmentPoints.push(point(tx, ty));
-
-//     t += 0.1;
-//   }
-
-//   // check the distance from line segments to the given point
-
-//   return false;
-// };
+/**
+ *
+ * @param p
+ * @param mid
+ * @param multiplier
+ * @returns
+ */
+export const pointScaleFromOrigin = <P extends GlobalPoint | LocalPoint>(
+  p: P,
+  mid: P,
+  multiplier: number,
+) => pointTranslate(mid, vectorScale(vectorFromPoint(p, mid), multiplier));
